pyproject.toml
[tool.poetry]
name = "ssd-api"
version = "0.1.0"
description = ""
authors = ["kkuga5121 <kkuga5121@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pendulum = "^3.0.0"
fastapi = "^0.114.1"
uvicorn = "^0.30.6"
pydub = "^0.25.1"
torch = "^2.4.1"
torchvision = "^0.19.1"
opencv-python = "^4.10.0.84"
pillow = "^10.4.0"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
